[Bugfix] QGIS Server / OGR loads schema when reading GML
When reading a GML file, OGR will load schemas defined in the `xsi:schemaLocation` attribute of `wfs:FeatureCollection` element.
In the context of QGIS Server, it means that the server wil request himself to get the specific layer schema. It can be block the request and generated a `504 gateway timeout Error`.
Schemas are not mandatory for reading and transforming GML file. To avoid blocking request, the solution is to remove the `xsi:schemaLocation` content.